--- a/src/config.rs
+++ b/src/config.rs
@@ -10,7 +10,7 @@
 use std::path::PathBuf;
 
 use base64;
-use sodiumoxide::crypto::sign::ed25519;
+use rust_sodium::crypto::sign::ed25519;
 
 use error::{Error, Result};
 use version::Version;