--- a/src/error.rs
+++ b/src/error.rs
@@ -25,6 +25,9 @@ pub enum Error {
     /// Public key in config on a given line could not be parsed as base64.
     InvalidPublicKeyData(usize, base64::DecodeError),
 
+    /// Secret key was not prefixed with "SECRET+".
+    InvalidSecretKeyPrefix,
+
     /// Secret key could not be parsed as base64, or the decoded key is invalid.
     InvalidSecretKeyData,