--- a/src/main.rs
+++ b/src/main.rs
@@ -20,6 +20,7 @@
 extern crate base64;
 extern crate filebuffer;
 extern crate ring;
+extern crate sodiumoxide;
 extern crate untrusted;
 
 use std::process;