--- a/src/manifest.rs
+++ b/src/manifest.rs
@@ -15,11 +15,8 @@ use std::path::{Path, PathBuf};
 use std::str;
 
 use base64;
-use ring::signature::Ed25519KeyPair;
-use ring::signature;
 use sodiumoxide::crypto::hash::sha256;
 use sodiumoxide::crypto::sign::ed25519;
-use untrusted::Input;
 
 use error::{Error, Result};
 use util;
@@ -352,11 +349,8 @@ pub fn store_local(path: &Path, bytes: &[u8]) -> Result<()> {
 
 #[cfg(test)]
 mod test {
-    use ring::signature::Ed25519KeyPair;
-    use ring::test::rand::FixedSliceRandom;
     use sodiumoxide::crypto::hash::sha256;
     use sodiumoxide::crypto::sign::ed25519;
-    use untrusted::Input;
 
     use error::Error;
     use super::{Entry, Manifest, parse_entry};