## Motivation

[Ring’s stability policy](https://github.com/briansmith/ring/blob/7e3682c02170f5eafa423b63bacfa018de2878a9/README.md#versioning--stability) is problematic; it explicitly does not support other Rust versions than the latest stable and it does not backport security fixes to older releases. This makes it impossible to securely use Ring on a platform that dictates an other Rust version, such as Ubuntu or Debian. (Ubuntu ships 1.25.0 at the time of writing while 1.28.0 is the most recent; Debian stable ships 1.24.1.)

Sodiumoxide on the other hand binds the most recent libsodium, and its latest release is compatible with Rust 1.24.1.

## User-visible Changes

Rather than using Ring’s format for storing key pairs and encoding that as base64, concatenate the private and public key bytes and use the base64 encoding of that as the secret key exposed to users. It is a bit longer than the ones used by Ring. Also prepend a `SECRET+` prefix, to make it harder to mistake the secret data for public data.

The keys themselves are still the same, only the formatting is different. The tests still use the same test key.

## Side Effects

Number of dependencies before: 28
Number of dependencies after:  11

Build time for a clean release build with `rustc 1.27.0-beta.7`:
Before: 26.40s, 25.16s, 25.50s
After:  12.68s, 12.40s, 12.45s

Binary size for a release build with `rustc 1.27.0-beta.7`, `opt-level` unset, linking libsodium statically:
Before 3666168 bytes (3.7 MB)
After: 5120784 bytes  (5.1 MB)

Binary size for a release build with `rustc 1.28.0 stable` with `opt-level="z"` (optimize for size), linking libsodium statically:
Before: 2438184 bytes (2.3 MB)
After:  3951544 bytes (4.0 MB)
Before, stripped: 584288 bytes (0.6 MB)
After, stripped: 957584 bytes (1.0 MB)

So we do get a significant regression in binary size, but also faster builds and less dependencies to worry about.

Fixes #19, although not with `ed25519-dalek` and `sha2`, but with `sodiumoxide` instead; `sha2` has a very good stability policy, but `ed25519-dalek` does not compile on Rust 1.24.1.