@ruuda 2018-08-04
Also @arianvp thanks for suggesting to look into what Wire uses! Sodiumoxide’s API is also nicer than Ring’s, I could get rid of some of the wrapper types.